Split out views and setting up for Detail view.
