+++ FetchRecipes/Views/RecipeListView.swift
@@ -24,7 +24,7 @@ struct RecipeListView: View {
         .padding()
       }
       .navigationTitle(model.listName)
-      .sheet(
+      .navigationDestination(
         unwrapping: $model.destination,
         case: /RecipeListModel.Destination.detail
       ) { $meal in
@@ -37,28 +37,8 @@ struct RecipeListView: View {
   }
 }
 
-struct MealCardView: View {
-  let meal: Meal
 
-  var body: some View {
-    VStack(alignment: .center) {
-      AsyncImage(url: meal.strMealThumb) { image in
-        image.resizable()
-      } placeholder: {
-        Color.gray.opacity(0.1)
-      }
-      .frame(width: 150, height: 150)
-      .cornerRadius(10)
-      .overlay(RoundedRectangle(cornerRadius: 10)
-        .stroke(Color.gray, lineWidth: 1))
-      .shadow(radius: 5)
 
-      Text(meal.strMeal)
-        .frame(maxWidth: .infinity, alignment: .center)
-        .lineLimit(1)
-    }
-  }
-}
 
 #Preview("Portrait") {
   RecipeListView(model: .mock(numberOfMeals: 5))
